% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coordinates.R
\name{create_coordinates}
\alias{create_coordinates}
\title{Function to create coordinates for a Hilbert Curve
This functions adds three columns to a data frame: reld, x, y}
\usage{
create_coordinates(df, idcol)
}
\arguments{
\item{df}{the dataframe to use}

\item{idcol}{the column to use for mapping}
}
\value{
a data frame with three additional columns
}
\description{
Function to create coordinates for a Hilbert Curve
This functions adds three columns to a data frame: reld, x, y
}
\examples{
mtcars \%>\% tibble::rownames_to_column() \%>\% create_id_column(rowname) \%>\% create_coordinates(gghid)
}
