% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id_column.R
\name{create_id_column}
\alias{create_id_column}
\title{Function to create keys from a character list}
\usage{
create_id_column(df, col)
}
\arguments{
\item{df}{the dataframe that is used}

\item{col}{the column name in NSE format}
}
\value{
a dataframe with an additional gghid column
}
\description{
Function to create keys from a character list
}
\examples{
mtcars \%>\% tibble::rownames_to_column() \%>\% create_id_column(rowname)
}
