% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d2xy}
\alias{d2xy}
\title{Returns the position for a distance in n}
\usage{
d2xy(n, d)
}
\arguments{
\item{n}{First value}

\item{d}{Second value}
}
\value{
Vector of x y
}
\description{
Returns the position for a distance in n
}
