% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d2xy2}
\alias{d2xy2}
\title{Returns the x/y-position for a Vector of distances d in n possible values}
\usage{
d2xy2(n, d)
}
\arguments{
\item{n}{Size of}

\item{d}{Second value}
}
\value{
Matrix of x y values
}
\description{
Returns the x/y-position for a Vector of distances d in n possible values
}
