% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghilbertstrings-package.R
\docType{package}
\name{gghilbertstrings-package}
\alias{gghilbertstrings}
\alias{gghilbertstrings-package}
\title{gghilbertstrings: A Fast 'ggplot2'-Based Implementation of Hilbert Curves}
\description{
A set of functions that help to create plots based on Hilbert curves. Hilbert curves are used to map one dimensional data into the 2D plane. The package provides a function that generate a 2D coordinate from an integer position. As a specific use case the package provides a function that allows mapping a character column in a data frame into 2D space using 'ggplot2'. This allows visually comparing long lists of URLs, words, genes or other data that has a fixed order and position.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sumidu/gghilbertstrings}
  \item Report bugs at \url{https://github.com/Sumidu/gghilbertstrings/issues}
}

}
\author{
\strong{Maintainer}: André Calero Valdez \email{andrecalerovaldez@gmail.com} (\href{https://orcid.org/0000-0002-6214-1461}{ORCID})

}
\keyword{internal}
