% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order4}
\alias{order4}
\title{Finds the order of the next highest number to the power of 4}
\usage{
order4(n)
}
\arguments{
\item{n}{number}
}
\value{
Order of next highest number 4^x
}
\description{
Finds the order of the next highest number to the power of 4
}
