% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghilbertstrings-package.R
\docType{package}
\name{gghilbertstrings-package}
\alias{gghilbertstrings}
\alias{gghilbertstrings-package}
\title{gghilbertstrings: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sumidu/gghilbertstrings}
  \item Report bugs at \url{https://github.com/Sumidu/gghilbertstrings/issues}
}

}
\author{
\strong{Maintainer}: André Calero Valdez \email{andrecalerovaldez@gmail.com} (\href{https://orcid.org/0000-0002-6214-1461}{ORCID})

}
\keyword{internal}
